requirements.txt
pillow
opencv-python
tensorflow==2.18.0
ultralytics